refactor: split select button into another component
