+++ src/components/Form/Select/types.ts
@@ -1,0 +1,4 @@
+export interface Option {
+  label: string;
+  value: string;
+}
